separate file for `LocalState`
